// src/models/produto-item/produto-item.interface.ts
export interface ProdutoItem {
    $key?: string;
    iCodigoFabricante: string;
    iCodigoInterno: string;
    iDescricao: string; //ok
    iLocalEstoque: string;
    iDescricaoReduzida: string; //ok
    iMarca: string; //ok
    iNumero: string; //ok
    iCor: string; //ok
    iColecao: string; //ok
    iLote: string;
    iTipo: string; //ok
    iGrupo: string; //ok
    iValor: string; 
    iDesconto: string; 
    iQtdEstoque: string; 
    iDataHoraCadastro: string;
}